create object module for better structure
