+++ pyformlang/finite_automaton/utils.py
@@ -1,39 +1,11 @@
 """ Utility for finite automata """
 
-from typing import Dict, List, AbstractSet, Iterable, Optional, Hashable
+from typing import Dict, List, AbstractSet, Iterable, Optional
 from numpy import empty
 
 from .state import State
 from .symbol import Symbol
 from .epsilon import Epsilon
-
-
-def to_state(given: Hashable) -> State:
-    """ Transforms the input into a state
-
-    Parameters
-    ----------
-    given : any
-        What we want to transform
-    """
-    if isinstance(given, State):
-        return given
-    return State(given)
-
-
-def to_symbol(given: Hashable) -> Symbol:
-    """ Transforms the input into a symbol
-
-    Parameters
-    ----------
-    given : any
-        What we want to transform
-    """
-    if isinstance(given, Symbol):
-        return given
-    if given in ("epsilon", "ɛ"):
-        return Epsilon()
-    return Symbol(given)
 
 
 def to_single_state(l_states: Iterable[State]) -> State:
