+++ pyformlang/objects/__init__.py
@@ -1,0 +1,13 @@
+""" Collection of object representations """
+
+import finite_automaton_objects
+import cfg_objects
+import regex_objects
+import pda_objects
+
+
+__all__ = ["finite_automaton_objects",
+           "cfg_objects",
+           "regex_objects",
+           "pda_objects"]
+
