+++ pyformlang/objects/cfg_objects/variable.py
@@ -1,7 +1,7 @@
 """ A variable in a CFG """
-import string
 
 from typing import Optional, Hashable
+from string import ascii_uppercase
 
 from .cfg_object import CFGObject
 
@@ -24,6 +24,6 @@ class Variable(CFGObject):  # pylint: disable=too-few-public-methods
 
     def to_text(self) -> str:
         text = str(self._value)
-        if text and text[0] not in string.ascii_uppercase:
+        if text and text[0] not in ascii_uppercase:
             return '"VAR:' + text + '"'
         return text
