+++ pyformlang/objects/finite_automaton_objects/epsilon.py
@@ -3,6 +3,7 @@ Represents an epsilon transition
 """
 
 from typing import Any
+
 from .symbol import Symbol
 
 
@@ -19,8 +20,8 @@ class Epsilon(Symbol):  # pylint: disable=too-few-public-methods
     def __init__(self) -> None:
         super().__init__("epsilon")
 
-    def __hash__(self) -> int:
-        return hash("EPSILON TRANSITION")
-
     def __eq__(self, other: Any) -> bool:
         return isinstance(other, Epsilon)
+
+    def __hash__(self) -> int:
+        return hash("EPSILON TRANSITION")
