+++ pyformlang/objects/pda_objects/stack_symbol.py
