+++ pyformlang/objects/regex_objects/regex_objects.py
@@ -5,8 +5,8 @@ Representation of some objects used in regex.
 from typing import List, Iterable
 from abc import abstractmethod
 
-from pyformlang.cfg import Production
-from pyformlang.cfg.utils import to_variable, to_terminal
+from cfg_objects.production import Production
+from cfg_objects.utils import to_variable, to_terminal
 
 
 class Node:  # pylint: disable=too-few-public-methods
@@ -56,38 +56,6 @@ class Node:  # pylint: disable=too-few-public-methods
         """ Gets the rules for a context-free grammar to represent the \
         operator"""
         raise NotImplementedError
-
-
-CONCATENATION_SYMBOLS = ["."]
-UNION_SYMBOLS = ["|", "+"]
-KLEENE_STAR_SYMBOLS = ["*"]
-EPSILON_SYMBOLS = ["epsilon", "$"]
-PARENTHESIS = ["(", ")"]
-
-SPECIAL_SYMBOLS = CONCATENATION_SYMBOLS + \
-                  UNION_SYMBOLS + \
-                  KLEENE_STAR_SYMBOLS + \
-                  EPSILON_SYMBOLS + \
-                  PARENTHESIS
-
-
-def to_node(value: str) -> Node:
-    """ Transforms a given value into a node """
-    if not value:
-        res = Empty()
-    elif value in CONCATENATION_SYMBOLS:
-        res = Concatenation()
-    elif value in UNION_SYMBOLS:
-        res = Union()
-    elif value in KLEENE_STAR_SYMBOLS:
-        res = KleeneStar()
-    elif value in EPSILON_SYMBOLS:
-        res = Epsilon()
-    elif value[0] == "\\":
-        res = Symbol(value[1:])
-    else:
-        res = Symbol(value)
-    return res
 
 
 class Operator(Node):  # pylint: disable=too-few-public-methods
