+++ pyformlang/regular_expression/__init__.py
@@ -21,7 +21,7 @@ Available Classes
 
 
 from .regex import Regex
-from .regex_objects import MisformedRegexError
+from ..objects.regex_objects.regex_objects import MisformedRegexError
 from .python_regex import PythonRegex
 
 __all__ = ["Regex", "PythonRegex", "MisformedRegexError"]
